move constant and regex out
