+++ azdev/operations/constant.py
